fix: fix coderabbit and number issues in multi modals
