+++ components/factory/modals/index.ts
@@ -1,4 +1,4 @@
 export * from './denomInfo';
 export * from './updateDenomMetadata';
-export * from './multiMfxMintModal';
-export * from './multiMfxBurnModal';
+export * from '../../admins/modals/multiMfxMintModal';
+export * from '../../admins/modals/multiMfxBurnModal';
